pyproject.toml
[tool.poetry]
name = "marl"
version = "0.1.0"
description = "Multi Agent Reinforcement Learning algorithms with the RLEnv library"
authors = ["Yannick Molinghen <yannick.molinghen@ulb.be>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
torch = "^1.13.1"
numpy = "^1.23.4"
rlenv = { git = "https://github.com/yamoling/rlenv.git" }
tensorboard = "^2.11.2"
opencv-python = "^4.7.0.68"
laser-env = { git = "https://github.com/raphaelavalos/laser-env.git" }
sumtree = "^0.2"
polars = "^0.16.18"
tqdm = "^4.65.0"
argcomplete = "^3.0.8"
pynvml = "^11.5.0"
pyserde = "^0.10.6"
lle = { git = "https://github.com/yamoling/lle" }

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.6.3"
pywebview = { extras = ["qt"], version = "^4.0.1" }
pycairo = "^1.23.0"
pygobject = "^3.42.2"
websockets = "^10.4"
flask = "^2.2.2"
flask-cors = "^3.0.10"
pytest = "^7.3.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["src"]


[tool.ruff]
fixable = ["ALL"]
line-length = 140
